+++ golang/.gitignore
@@ -1,0 +1,3 @@
+cmd/sig0namectl/sig0namectl
+sig0namectl
+K*.private
