+++ golang/wasm/README.md
@@ -1,0 +1,4 @@
+Links:
+
+* https://golangbot.com/webassembly-using-go/
+
